Add support for colspan & rowspan
* Add test cases
* Update conversionScriptPart to support the colspan and rowspan properties
* Conversion.js has a col counter now that differs from the i/j counter
used to loop through the table.
* This curr_col should be used to set the column value to the output table.
* rowspan support differs that an array of objects need to store row/col
locations that have been 'spanned'.  Thus on each write of a row, the
writer needs to check if a column should be skipped because a previous
write to that column.  Once the reference to the rowspan has expired as the
writer has moved past it, it can be deleted from the array.

Add rowspan support to conversion.js